+++ classes/ui/wpdk-listtable-viewcontroller.php
@@ -487,26 +487,23 @@ class WPDKListTableViewController extends WP_List_Table {
 
       <?php $this->search_box_field() ?>
 
-      <?php if ( isset( $_REQUEST['page'] ) ) : ?>
-        <input type="hidden" name="page" value="<?php echo $_REQUEST['page'] ?>" />
-      <?php endif; ?>
-
-      <?php if ( isset( $_REQUEST['post_type'] ) ) : ?>
-        <input type="hidden" name="post_type" value="<?php echo $_REQUEST['post_type'] ?>" />
-      <?php endif; ?>
-
-      <?php if ( isset( $_REQUEST['orderby'] ) ) : ?>
-        <input type="hidden" name="orderby" value="<?php echo $_REQUEST['orderby'] ?>" />
-      <?php endif; ?>
-
-      <?php if ( isset( $_REQUEST['order'] ) ) : ?>
-        <input type="hidden" name="order" value="<?php echo $_REQUEST['order'] ?>" />
-      <?php endif; ?>
+      <?php echo $this->html_filters() ?>
 
       <?php do_action( 'wpdk_list_table_form', $this ); // @deprecated action since 1.5.1 - use 'before_display()' instead ?>
 
       <?php unset( $_REQUEST['action'] ); ?>
       <?php $_SERVER['REQUEST_URI'] = isset( $_REQUEST['_wp_http_referer'] ) ? $_REQUEST['_wp_http_referer'] : $_SERVER['REQUEST_URI'] ?>
+
+      <?php
+      $filters = $this->get_filters();
+      $filter_args = array();
+      foreach ( $filters as $key => $value ) {
+        if ( isset( $_REQUEST[ $key ] ) && !empty( $_REQUEST[ $key ] )) {
+          $filter_args[ $key ] = urlencode( $_REQUEST[ $key ] );
+        }
+      }
+      $_SERVER['REQUEST_URI'] = add_query_arg( $filter_args, $_SERVER['REQUEST_URI'] );
+      ?>
 
       <?php $this->before_display(); // since 1.5.1 ?>
 
@@ -525,6 +522,48 @@ class WPDKListTableViewController extends WP_List_Table {
 
     return $this->viewController->html();
   }
+
+  /**
+   * Return a set of registered filters
+   *
+   * @brief Brief
+   */
+  protected function get_filters()
+  {
+    $standard_filters = array(
+      'page'      => array(),
+      'post_type' => array(),
+      'orderby'   => array(),
+      'order'     => array(),
+    );
+
+    if ( !empty( $this->model ) && method_exists( $this->model, 'get_filters' ) ) {
+      $standard_filters = array_merge( $standard_filters, (array)$this->model->get_filters() );
+    }
+
+    return $standard_filters;
+  }
+
+  /**
+   * Return a set of input hidden fields for registered filters
+   *
+   * @brief Brief
+   *
+   * @return string
+   */
+  protected function html_filters()
+  {
+    WPDKHTML::startCompress();
+    foreach ( $this->get_filters() as $request => $value ) {
+      if ( isset( $_REQUEST[ $request ] ) && !empty( $_REQUEST[ $request ] ) ) :
+        ?><input type="hidden" name="<?php echo $request ?>" value="<?php echo urlencode( $_REQUEST[ $request ] ) ?>" /><?php
+      endif;
+    }
+
+    return WPDKHTML::endCompress();
+  }
+
+
 
   /**
    * Called when the title has been drawed
@@ -840,7 +879,8 @@ class WPDKListTableViewController extends WP_List_Table {
   public function column_cb( $item )
   {
     $name  = $this->args['singular'];
-    $value = $item[$name];
+    $value = $item[ $name ];
+
     return sprintf( '<input type="checkbox" name="%s[]" value="%s" />', $name, $value );
   }
 
@@ -1078,6 +1118,19 @@ class WPDKListTableModel {
   }
 
   /**
+   * Return a key values array with registered filters
+   *
+   * @brief Filters
+   * @since 1.5.2
+   *
+   * @return array
+   */
+  public function get_filters()
+  {
+    return array();
+  }
+
+  /**
    * Return a key value pairs array with the list of columns
    *
    * @brief Return the list of columns
@@ -1218,6 +1271,20 @@ class WPDKListTableModel {
           'action2'         => false,
           'page'            => isset( $_REQUEST['page'] ) ? $_REQUEST['page'] : false,
         );
+
+        // Previous selected filters
+        $filters = $this->get_filters();
+        $filter_args = array();
+        foreach ( $filters as $key => $value ) {
+          if ( isset( $_REQUEST[ $key ] ) && !empty( $_REQUEST[ $key ] )) {
+            $filter_args[ $key ] = urlencode( $_REQUEST[ $key ] );
+          }
+        }
+
+        //  merge standard args with filters args
+        $args = array_merge( $args, $filter_args );
+
+        // New referrer
         $uri  = add_query_arg( $args, $_REQUEST['_wp_http_referer'] );
 
         wp_safe_redirect( $uri );